Add DateTimeValueTests and refactor for compatibility
